delete api

+++ src/main/java/com/example/lms/application/entity/Application.java
@@ -17,12 +17,12 @@ public class Application {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private Long lectureId;
+    private Integer id;
+    private Integer lectureId;
     private String lectureName;
     private String professorName;
     private Integer score;
-    private Long maximumNumber;
+    private Integer maximumNumber;
     private Integer startTime;
     private DayOfWeek dayOfWeek;
     private Semester semester;
